# documentation/postman/buildingexaminator/v1/openapi.yaml
openapi: 3.0.0
info:
  title: 'Building Examinator'
  version: 0.0.1
paths:
  /housingstocks:
    get:
      summary: 'Returns details about multiple housing stocks'
      operationId: 19e582c4a577484d38858cd48abd01ae
      responses:
        '200':
          description: 'Details about multiple housing stocks'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/housingStocks'
  '/housingstocks/{housingStockId}':
    get:
      summary: 'Returns details about a housing stock'
      operationId: 690081229387f635c736db5dc1b4e69a
      parameters:
        -
          name: housingStockId
          in: path
          description: 'The id of the housing stock'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Details about a housing stock'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HousingStock'
  '/housingstocks/{housingStockId}/blocks':
    get:
      summary: 'Returns details about multiple blocks'
      operationId: 640b4429ebb1462f1d1f14c965a9b2bb
      parameters:
        -
          name: housingStockId
          in: path
          description: 'The id of the housing stock'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Details about multiple blocks'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/blocks'
  '/housingstocks/{housingStockId}/blocks/{blockId}':
    get:
      summary: 'Returns details about a block'
      operationId: 7048b6f03a7bb1deb3f4f1cc26c1d2e5
      parameters:
        -
          name: housingStockId
          in: path
          description: 'The id of the housing stock'
          required: true
          schema:
            type: integer
            format: int64
        -
          name: blockId
          in: path
          description: 'The id of a block'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Details about a block'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Block'
  '/housingstocks/{housingStockId}/addresses':
    get:
      summary: 'Returns details about multiple addresses'
      operationId: 7bdb42033e6379ec7cbd3d69c82246eb
      parameters:
        -
          name: housingStockId
          in: path
          description: 'The id of the housing stock'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Details about multiple addresses'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/buildingAddresses'
  '/housingstocks/{housingStockId}/addresses/{addressId}':
    get:
      summary: 'Returns details about an address'
      operationId: 07133fa35032ffada708a0acd50e94f7
      parameters:
        -
          name: housingStockId
          in: path
          description: 'The id of the housing stock'
          required: true
          schema:
            type: integer
            format: int64
        -
          name: addressId
          in: path
          description: 'The id of an address'
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 'Details about an address'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BuildingAddress'
components:
  schemas:
    housingStocks:
      title: 'Housing stocks'
      description: 'An array of housing stocks'
      type: array
      items:
        $ref: '#/components/schemas/HousingStock'
    blocks:
      title: Blocks
      description: 'An array of blocks'
      type: array
      items:
        $ref: '#/components/schemas/Block'
    buildingAddresses:
      title: Addresses
      description: 'An array of addresses'
      type: array
      items:
        $ref: '#/components/schemas/BuildingAddress'
    Block:
      properties:
        numberOfBuildingAddresses:
          type: integer
        buildingTypeSelection: {  }
        financialNumber:
          type: string
        code:
          type: string
        name:
          type: string
        description:
          type: string
        creationTime:
          type: string
          format: date-time
        lastChangeTime:
          type: string
          format: date-time
        id:
          type: integer
      type: object
    BuildingAddress:
      properties:
        residentialArea:
          $ref: '#/components/schemas/ResidentialArea'
        buildingType:
          $ref: '#/components/schemas/BuildingType'
        livingType:
          $ref: '#/components/schemas/LivingType'
        blocks: {  }
        rentalUnitNumber:
          type: string
        streetName:
          type: string
        houseNumber:
          type: integer
        addition:
          type: string
        zipcode:
          type: string
        city:
          type: string
        bagId:
          type: integer
        constructionYear:
          type: integer
        renovationYear:
          type: integer
        creationTime:
          type: string
          format: date-time
        lastChangeTime:
          type: string
          format: date-time
        id:
          type: integer
      type: object
    BuildingType:
      properties:
        buildingAddresses: {  }
        buildingTypeSelections: {  }
        code:
          type: string
        name:
          type: string
        description:
          type: string
        creationTime:
          type: string
          format: date-time
        lastChangeTime:
          type: string
          format: date-time
        id:
          type: integer
      type: object
    BuildingTypeSelection:
      properties:
        code:
          type: string
        name:
          type: string
        description:
          type: string
        creationTime:
          type: string
          format: date-time
        lastChangeTime:
          type: string
          format: date-time
        id:
          type: integer
    HousingStock:
      properties:
        blocks: {  }
        numberOfBlocks:
          type: integer
        numberOfBuildingAddresses:
          type: integer
        code:
          type: string
        name:
          type: string
        description:
          type: string
        creationTime:
          type: string
          format: date-time
        lastChangeTime:
          type: string
          format: date-time
        id:
          type: integer
      type: object
    LivingType:
      properties:
        buildingAddresses: {  }
        code:
          type: string
        name:
          type: string
        description:
          type: string
        creationTime:
          type: string
          format: date-time
        lastChangeTime:
          type: string
          format: date-time
        id:
          type: integer
      type: object
    ResidentialArea:
      properties:
        buildingAddresses: {  }
        code:
          type: string
        name:
          type: string
        description:
          type: string
        creationTime:
          type: string
          format: date-time
        lastChangeTime:
          type: string
          format: date-time
        id:
          type: integer
      type: object